Restructure and remove duplication
