+++ src/components/Charts/types.tsx
@@ -1,0 +1,12 @@
+export interface Margin {
+  top?: number;
+  bottom?: number;
+  right?: number;
+  left?: number;
+}
+
+export interface Level {
+  level: string;
+  start: number;
+  end: number;
+}
