+++ src/components/Charts/utils.tsx
@@ -1,0 +1,12 @@
+export function levelColor(level: string) {
+  switch (level) {
+    case "high":
+      return "#3baa34";
+    case "mid":
+      return "#fd9c00";
+    case "low":
+      return "#e30713";
+    default:
+      throw new Error(`${level} is not a valid level`);
+  }
+}
